feat: Implement Explore Card and Sync Card components
- Added ExploreCardData model to encapsulate data for explore cards.
- Created ExploreCard widget to display course, collection, and content cards with tags, metadata, and actions.
- Introduced SyncProvider for managing download and upload operations with state management.
- Developed TransferCardData model to represent transfer operations with progress tracking.
- Implemented TransferCard widget for displaying transfer details, including actions for pause, resume, cancel, and delete.
- Added DownloadsTabView and UploadsTabView for managing download and upload views respectively.
- Enhanced ScaleClickWrapper for animated button interactions.

+++ lib/features/browse/shared/usecases/collections/modify_collection_uc.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'dart:math' as math;
 
 import 'package:slidesync/core/storage/hive_data/app_hive_data.dart';
@@ -20,8 +19,8 @@ class ModifyCollectionUc {
         final filePath = item.path.filePath;
         await FileUtils.deleteFileAtPath(filePath);
 
-        final previewPath = FileDetails.fromMap(item.metadata.thumbnails ?? {}).filePath;
-        await FileUtils.deleteFileAtPath(previewPath);
+        final previewPath = item.metadata.thumbnails?.filePath;
+        if (previewPath != null && previewPath.isNotEmpty) await FileUtils.deleteFileAtPath(previewPath);
         fileSum += item.fileSize;
       }
     }
